requirements.txt
beautifulsoup4
fastapi
pydantic
Requests
uvicorn
lxml
sqlalchemy
httpx
bs4
tortoise-orm